% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_packb.R
\name{packb_f1}
\alias{packb_f1}
\title{f1 exposed by packb}
\usage{
packb_f1()
}
\value{
A String "Hello from packb_f1"
}
\description{
This is function \code{f1} exposed by \code{packb}
}
\examples{
packb_f1()
}

